+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Update design of `Spinner`.
 * Add `InputSpinner` which is used in buttons and input fields.
 * Merge the two individual theme prop for calendar and text input into one common theme prop for `DateTextInput`.
+* Add `fontWeight` to `ButtonTheme`.
 
 #### 0.0.33